scripts/requirements.txt
yfinance
requests
beautifulsoup4
google-genai
pytz
pandas
numpy
feedparser
